// packages/sui-grpc/src/proto/sui/rpc/v2beta/executed_transaction.ts
export interface ExecutedTransaction {
	/**
	 * The digest of this Transaction.
	 *
	 * @generated from protobuf field: optional string digest = 1;
	 */
	digest?: string;
	/**
	 * The transaction itself.
	 *
	 * @generated from protobuf field: optional sui.rpc.v2beta.Transaction transaction = 2;
	 */
	transaction?: Transaction;
	/**
	 * List of user signatures that are used to authorize the
	 * execution of this transaction.
	 *
	 * @generated from protobuf field: repeated sui.rpc.v2beta.UserSignature signatures = 3;
	 */
	signatures: UserSignature[];
	/**
	 * The `TransactionEffects` for this transaction.
	 *
	 * @generated from protobuf field: optional sui.rpc.v2beta.TransactionEffects effects = 4;
	 */
	effects?: TransactionEffects;
	/**
	 * The `TransactionEvents` for this transaction.
	 *
	 * This field might be empty, even if it was explicitly requested, if the
	 * transaction didn't produce any events.
	 * `sui.types.TransactionEffects.events_digest` is populated if the
	 * transaction produced any events.
	 *
	 * @generated from protobuf field: optional sui.rpc.v2beta.TransactionEvents events = 5;
	 */
	events?: TransactionEvents;
	/**
	 * The sequence number for the checkpoint that includes this transaction.
	 *
	 * @generated from protobuf field: optional uint64 checkpoint = 6;
	 */
	checkpoint?: bigint;
	/**
	 * The Unix timestamp of the checkpoint that includes this transaction.
	 *
	 * @generated from protobuf field: optional google.protobuf.Timestamp timestamp = 7;
	 */
	timestamp?: Timestamp;
	/**
	 * @generated from protobuf field: repeated sui.rpc.v2beta.BalanceChange balance_changes = 8;
	 */
	balanceChanges: BalanceChange[];
	/**
	 * Set of input objects used during the execution of this transaction.
	 *
	 * @generated from protobuf field: repeated sui.rpc.v2beta.Object input_objects = 10;
	 */
	inputObjects: Object[];
	/**
	 * Set of output objects produced from the execution of this transaction.
	 *
	 * @generated from protobuf field: repeated sui.rpc.v2beta.Object output_objects = 11;
	 */
	outputObjects: Object[];
}
// ...
export const ExecutedTransaction = new ExecutedTransaction$Type();